Fix field type of StreamSourceFieldMapper

+++ inlong-manager/manager-dao/src/main/java/org/apache/inlong/manager/dao/entity/StreamSourceFieldEntity.java
@@ -43,11 +43,11 @@ public class StreamSourceFieldEntity implements Serializable {
 
     private String fieldComment;
 
-    private Short isMetaField;
+    private Integer isMetaField;
 
     private String fieldFormat;
 
-    private Short rankNum;
+    private Integer rankNum;
 
     private Integer isDeleted;
 }
